backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── Supabase ──────────────────────────────────────────────────────────────────
supabase>=2.4.0

# ── Environment & config ────────────────────────────────────────────────────
python-dotenv>=1.0.0
python-multipart>=0.0.9   # required for FastAPI file uploads / Form fields

# ── HTTP ────────────────────────────────────────────────────────────────────
httpx>=0.27.0

# ── LangChain / Agent pipeline ────────────────────────────────────────────────
langchain-core>=0.2.0
langchain-anthropic>=0.1.0
langsmith>=0.1.0
anthropic>=0.84.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=9.0.0
pytest-asyncio>=1.3.0
anyio>=4.0.0
